Update qtwebkit to branch 5.8
This updates qtwebkit to the 5.8 branch on
https://github.com/qt/qtwebkit/commits/5.8.  The specific hash for this
version of the branch is 0b48569e2bf9afc1b6ca5e359c3a948dd8c77619.

This is the only source code for version 5.8 that I have been able to
find.  There is no tag for it yet, so perhaps they are still updating it.